refactor: 패키지 이름을 bank에서 payment로 변경
