--- src/main/java/aegis/server/domain/bank/service/parser/TransactionParser.java
@@ -1,9 +1,0 @@
-package aegis.server.domain.bank.service.parser;
-
-import aegis.server.domain.bank.domain.Transaction;
-
-public interface TransactionParser {
-
-    Transaction parse(String transactionLog);
-    
-}
