+++ src/main/java/aegis/server/domain/payment/service/parser/TransactionParser.java
@@ -1,0 +1,9 @@
+package aegis.server.domain.payment.service.parser;
+
+import aegis.server.domain.payment.domain.Transaction;
+
+public interface TransactionParser {
+
+    Transaction parse(String transactionLog);
+
+}
