+++ src/main/java/aegis/server/global/config/BankConfig.java
@@ -1,7 +1,7 @@
 package aegis.server.global.config;
 
-import aegis.server.domain.bank.service.parser.IbkTransactionParser;
-import aegis.server.domain.bank.service.parser.TransactionParser;
+import aegis.server.domain.payment.service.parser.IbkTransactionParser;
+import aegis.server.domain.payment.service.parser.TransactionParser;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
